learn-english-from-famous-project/pyproject.toml
[project]
name = "learn-english-from-famous-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "dashscope>=1.25.13",
    "fastapi>=0.135.1",
    "jinja2>=3.1.6",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.22",
    "requests>=2.32.5",
    "uvicorn[standard]>=0.41.0",
]
